+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Removed
+
+- [TD-8054] Remove enriched_text field type
+
 ## [8.4.2] 2026-04-06
 
 ### Changed